irq: Support multiple interrupt ranges
In order to create a predictable interrupt map, each type of device gets
an interrupt range. This way an external DTB/ACPI table generator
knows about IRQ routing without having to guess in which order various
devices ask for IRQs.

This is needed for Arm CCA, where a client produces a DTB to be loaded
as-is into guest memory, so that it can be measured and attested. On Arm
PCI will get a fixed interrupt range, as well as virtio-mmio and the
UART.

Since an external software doesn't necessarily about detailed VFIO
capabilities, allocate an IRQ line for each VFIO device even if it's not
used. As a result the association between BDF and IRQ line is
well-defined and can be statically described in firmware tables.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>